+++ changelog.md
@@ -33,6 +33,7 @@
 * [2416](https://github.com/zeta-chain/node/pull/2416) - add Solana chain information
 * [2465](https://github.com/zeta-chain/node/pull/2465) - add Solana inbound SOL token observation
 * [2497](https://github.com/zeta-chain/node/pull/2416) - support for runtime chain (de)provisioning
+* [2518](https://github.com/zeta-chain/node/pull/2518) - add support for Solana address in zetacore
 
 ### Refactor